Add type castings and fix variables names

+++ src/proto/GameClient.cpp
@@ -82,7 +82,7 @@ namespace proto {
 
     unsigned long GameClient::newGame(unsigned short width, unsigned short height)
     {
-        NewGameRequest request(width, height);
+        NewGameRequest request(static_cast<uint8_t>(width), static_cast<uint8_t>(height));
         auto response = exchangeMessage<NewGameRequest, NewGameResponse>(request);
         return response.gameId;
     }
@@ -171,7 +171,7 @@ namespace proto {
         if (gameOver)
             throw ProtocolError(GAME_OVER);
 
-        TurnRequest request(gameId, playerId, target.getX(), target.getY());
+        TurnRequest request(gameId, playerId, static_cast<uint8_t>(target.getX()), static_cast<uint8_t>(target.getY()));
         auto response = exchangeMessage<TurnRequest, TurnResponse>(request);
         won = response.won;
         return response.hitResult;
